Fix breaking tests

+++ blockapps-haskell/bloc/bloc22/server/test/contracts/SimpleStorageGetSource.sol
@@ -13,12 +13,12 @@ contract SimpleStorage {
         return "contract SimpleStorage {\n  uint storedData;\n\n  function SimpleStorage() {\n    storedData = 1;\n  }\n\n  function set(uint x) {\n    storedData = x;\n  }\n\n  function get() constant returns (uint) {\n    return storedData;\n  }\n}\n";  
     
     }
-    function get() constant public returns (uint256) {
+    function get() constant public returns (uint) {
         return storedData;
   
     
     }
-    function set(uint256 x) public {
+    function set(uint x) public {
         storedData = x;
